Compute eigvalues and eigvectors

Return eigvalues and eigvectors of a matrix

Usage

get_eigen(X)

get_weighted_eigen(X)

Arguments

X

X_active

Details

Standardization depends on what you need to perform factor analysis. We implemented two types:

get_weighted_eigen: This is the default method in FactoMineR to compute eigvalues, eigvectors.

get_eigen: This is the standard method to compute eigvalues, eigvectors.

Examples

library(FactoMineR2)

iris[, -5] |>
  standardize_norm() |>
  get_eigen()
#> eigen() decomposition
#> $values
#> [1] 2.91849782 0.91403047 0.14675688 0.02071484
#> 
#> $vectors
#>                   Dim.1       Dim.2      Dim.3      Dim.4
#> Sepal.Length  0.5210659 -0.37741762  0.7195664  0.2612863
#> Sepal.Width  -0.2693474 -0.92329566 -0.2443818 -0.1235096
#> Petal.Length  0.5804131 -0.02449161 -0.1421264 -0.8014492
#> Petal.Width   0.5648565 -0.06694199 -0.6342727  0.5235971
#>